moved dbconnector to Factories and deleted example.php
