--- src/example.php
@@ -1,1 +1,0 @@
-<?php
